requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
pydantic>=2.10.0
pydantic-settings>=2.6.0
hidapi>=0.14.0

# Testing
pytest>=8.0.0
httpx>=0.27.0

# Type checking
mypy>=1.11.0
